% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{The Birthday Function}
\usage{
birthday(x)
}
\arguments{
\item{x}{the number of people within the group}
}
\value{
a probability scalar
}
\description{
The Birthday Function
}
\examples{
birthday(x = 20:25)
}
